fix: return 404s on forbidden access of resources

+++ backend/internal/db/queries/goals.sql
@@ -6,14 +6,14 @@ RETURNING *;
 -- name: UpdateGoalStatus :one
 UPDATE goals
 SET status = $1
-WHERE id = $2
+WHERE id = $2 AND user_id = $3
 RETURNING *;
 
 -- name: GetGoalsByUserId :many
 SELECT * FROM goals WHERE user_id = $1;
 
 -- name: GetGoalById :one
-SELECT * FROM goals WHERE id = $1 LIMIT 1;
+SELECT * FROM goals WHERE id = $1 AND user_id = $2 LIMIT 1;
 
 -- name: UpdateGoalById :one
 UPDATE goals
@@ -21,8 +21,8 @@ SET title = coalesce(sqlc.narg('title'), title),
     description = coalesce(sqlc.narg('description'), description),
     status = coalesce(sqlc.narg('status'), status),
     category_id = coalesce(sqlc.narg('category_id'), category_id)
-WHERE id = sqlc.arg('id')
+WHERE id = sqlc.arg('id') AND user_id = sqlc.arg('user_id')
 RETURNING *;
 
--- name: DeleteGoalById :exec
-DELETE FROM goals WHERE id = $1;
+-- name: DeleteGoalById :execrows
+DELETE FROM goals WHERE id = $1 AND user_id = $2;
